+++ infrastructure/Repositories/SocialRepository.php
@@ -8,7 +8,7 @@ use Domain\Entities\{
 use Infrastructure\DataSources\Database\{
     Users, UsersNickName, UsersStatus, SocialAccounts
 };
-use Infrastructure\Interfaces\SocialRepositoryInterface;
+use Infrastructure\Interfaces\Auth\SocialRepositoryInterface;
 use Laravel\Socialite\Contracts\User as SocialUser;
 use stdClass;
